Fix some evaluation bugs
This commit will add a check if first input is empty when an operator is called.
This check prevents a certain scenario from happening that sets the
first number to NaN.
Another check is added if there is no second operand provided when an
operator is called. The second operand is set to equal
to the first operand and the operator evaluates if equals is called.
The current operator changes if another operator is called while
there is no second operand provided.